Code refactoring.

+++ src/main/java/com/mrfisherman/relice/Service/Asset/AssetServiceImpl.java
@@ -1,14 +1,20 @@
 package com.mrfisherman.relice.Service.Asset;
 
 import com.mrfisherman.relice.Dto.AssetDto;
+import com.mrfisherman.relice.Dto.FloorDto;
 import com.mrfisherman.relice.Entity.Asset.AssetEntity;
 import com.mrfisherman.relice.Repository.AssetRepository;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.ui.Model;
+import org.springframework.web.server.ResponseStatusException;
+
 import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Set;
+import java.util.stream.Collectors;
 
 @Service
 public class AssetServiceImpl implements AssetService {
@@ -16,9 +22,9 @@ public class AssetServiceImpl implements AssetService {
     private final AssetRepository assetRepository;
     private final ModelMapper modelMapper;
 
-    public AssetServiceImpl(AssetRepository assetRepository) {
+    public AssetServiceImpl(AssetRepository assetRepository, ModelMapper modelMapper) {
         this.assetRepository = assetRepository;
-        this.modelMapper = new ModelMapper();
+        this.modelMapper = modelMapper;
     }
 
     @Override
@@ -27,8 +33,20 @@ public class AssetServiceImpl implements AssetService {
     }
 
     @Override
-    public void deleteAsset(Long id) throws IllegalArgumentException {
-        assetRepository.deleteById(id);
+    public List<Long> saveAssets(List<AssetDto> assets) {
+        return assets
+                .stream()
+                .map(x -> assetRepository.save(modelMapper.map(x, AssetEntity.class)).getId())
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    public void deleteAsset(Long id) throws EntityNotFoundException {
+        if (assetRepository.existsById(id)) {
+            assetRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException("No asset with id: " + id);
+        }
     }
 
     @Override
@@ -37,7 +55,7 @@ public class AssetServiceImpl implements AssetService {
         if (assetRepository.findById(id).isPresent()) {
             asset = modelMapper.map(assetRepository.findById(id), AssetDto.class);
         } else {
-            throw new EntityNotFoundException("No user with id: " + id);
+            throw new EntityNotFoundException("No asset with id: " + id);
         }
         return asset;
     }
@@ -52,5 +70,13 @@ public class AssetServiceImpl implements AssetService {
     public List<AssetDto> findAllAssets() {
         List<AssetEntity> assets = assetRepository.findAllWithoutNPlusOne();
         return modelMapper.map(assets, new TypeToken<List<AssetEntity>>() {}.getType());
+    }
+
+    public void updateAsset(AssetDto updatedAsset) {
+        if (updatedAsset != null) {
+            saveAsset(updatedAsset);
+        } else {
+            throw new IllegalArgumentException();
+        }
     }
 }
